this was harder than it looked
